Version 2.4
commit eab751215c485185dec96f4d0fbebc7dff7d898f
Author: Nik Reiman <nik.reiman@gmail.com>
Date:   Sat Dec 21 18:20:49 2013 +0100

    Fix syntax highlighting in README

    Fixes issue #9

commit 1f137a33e4d4357a2818aa8c1c6b6726dc0ffa83
Author: Nik Reiman <nik.reiman@gmail.com>
Date:   Fri Dec 20 18:03:49 2013 +0100

    Add newlines

commit ff5d740e183918c53950e25d9e3f2adf3f5239fb
Author: Nik Reiman <nik.reiman@gmail.com>
Date:   Fri Dec 20 18:03:20 2013 +0100

    DecibelParameter is now logarithmic

    Fixes issue #10

commit c012979fc72a0003cdc5fa95b328f5932624b616
Author: Nik Reiman <nik.reiman@gmail.com>
Date:   Fri Dec 20 18:03:01 2013 +0100

    Nicer display text for some parameter types

commit 9f63dfaf84d00c42e026d93ebca61ce0eff3a019
Author: Nik Reiman <nik.reiman@gmail.com>
Date:   Mon Dec 16 21:25:38 2013 +0100

    Update submodule to include Mac compilation fix

commit d962f9aab49c0679c54bf354a6775c152b5933bf
Author: Nik Reiman <nik.reiman@gmail.com>
Date:   Mon Dec 16 09:47:32 2013 +0100

    Add build info with regards to C++11 support

+++ include/IntegerParameter.h
@@ -36,11 +36,12 @@ public:
   int inMinValue, int inMaxValue, int inDefaultValue) :
   FloatParameter(inName, (ParameterValue)inMinValue,
     (ParameterValue)inMaxValue, (ParameterValue)inDefaultValue) {}
+
   virtual ~IntegerParameter() {}
 
   virtual const ParameterString getDisplayText() const {
     std::stringstream numberFormatter;
-    numberFormatter << getValue();
+    numberFormatter << (int)getValue();
     std::string result = numberFormatter.str();
     if(getUnit().length() > 0) {
       result.append(" ").append(getUnit());
